requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-database-url==1.0.3
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
